feat: add view containers id into elements
