+++ packages/core-browser/src/layout/view-id.ts
@@ -1,0 +1,14 @@
+const VIEW_CONTAINERS_PREFIX = 'opensumi';
+
+export namespace VIEW_CONTAINERS {
+  export const LEFT_TABBAR_PANEL = `${VIEW_CONTAINERS_PREFIX}-left-tabbar-panel`;
+  export const LEFT_TABBAR = `${VIEW_CONTAINERS_PREFIX}-left-tabbar`;
+  export const RIGHT_TABBAR_PANEL = `${VIEW_CONTAINERS_PREFIX}-right-tabbar-panel`;
+  export const RIGHT_TABBAR = `${VIEW_CONTAINERS_PREFIX}-right-tabbar`;
+  export const BOTTOM_TABBAR = `${VIEW_CONTAINERS_PREFIX}-bottom-tabbar`;
+  export const BOTTOM_TABBAR_PANEL = `${VIEW_CONTAINERS_PREFIX}-bottom-tabbar-panel`;
+  export const STATUS_BAR = `${VIEW_CONTAINERS_PREFIX}-status-bar`;
+  export const MENU_BAR = `${VIEW_CONTAINERS_PREFIX}-menu-bar`;
+  export const EDITOR = `${VIEW_CONTAINERS_PREFIX}-editor`;
+  export const EDITOR_TABS = `${VIEW_CONTAINERS_PREFIX}-editor-tabs`;
+}
