+++ packages/editor/src/browser/tab.view.tsx
@@ -18,6 +18,7 @@ import {
 import { InlineActionBar } from '@opensumi/ide-core-browser/lib/components/actions';
 import { Scroll } from '@opensumi/ide-core-browser/lib/components/scroll';
 import { LAYOUT_VIEW_SIZE } from '@opensumi/ide-core-browser/lib/layout/constants';
+import { VIEW_CONTAINERS } from '@opensumi/ide-core-browser/lib/layout/view-id';
 import { IMenuRegistry, MenuId } from '@opensumi/ide-core-browser/lib/menu/next';
 import { useInjectable, useUpdateOnEventBusEvent } from '@opensumi/ide-core-browser/lib/react-hooks';
 
@@ -384,7 +385,7 @@ export const Tabs = ({ group }: ITabsProps) => {
   );
 
   return (
-    <div className={styles.kt_editor_tabs}>
+    <div id={VIEW_CONTAINERS.EDITOR_TABS} className={styles.kt_editor_tabs}>
       <div className={styles.kt_editor_tabs_scroll_wrapper}>
         {!wrapMode ? (
           <Scroll ref={(el) => (el ? (tabContainer.current = el.ref) : null)} className={styles.kt_editor_tabs_scroll}>
