+++ packages/main-layout/src/browser/tabbar/styles.module.less
@@ -244,8 +244,8 @@
     }
   }
 }
-.kt_left_tab,
-.kt_right_tab {
+.left_tab,
+.right_tab {
   margin-top: 8px;
   background-color: var(--activityBar-background);
   width: 48px;
@@ -272,7 +272,7 @@
   }
 }
 
-.kt_left_tab {
+.left_tab {
   &:global(.active) {
     .icon_tab {
       &::before {
@@ -283,7 +283,7 @@
   }
 }
 
-.kt_right_tab {
+.right_tab {
   &:global(.active) {
     .icon_tab {
       &::before {
@@ -294,7 +294,7 @@
   }
 }
 
-.kt_bottom_tab {
+.bottom_tab {
   margin: 0 4px;
   padding: 0 8px;
   position: relative;
