+++ packages/startup/entry/sample-modules/menu-bar/menu-bar.view.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react';
 
 import { LAYOUT_VIEW_SIZE } from '@opensumi/ide-core-browser/lib/layout/constants';
+import { VIEW_CONTAINERS } from '@opensumi/ide-core-browser/lib/layout/view-id';
 import { MenuBar } from '@opensumi/ide-menu-bar/lib/browser/menu-bar.view';
 
 import * as styles from './menu-bar.module.less';
@@ -11,7 +12,7 @@ import * as styles from './menu-bar.module.less';
  * opensumi's original menubar.
  */
 export const MenuBarView = () => (
-  <div className={styles.menu_bar_view} style={{ height: LAYOUT_VIEW_SIZE.MENUBAR_HEIGHT }}>
+  <div id={VIEW_CONTAINERS.MENU_BAR} className={styles.menu_bar_view} style={{ height: LAYOUT_VIEW_SIZE.MENUBAR_HEIGHT }}>
     <span className={styles.menu_bar_logo} />
     <MenuBar />
   </div>
