+++ packages/status-bar/src/browser/status-bar.view.tsx
@@ -2,13 +2,13 @@ import cls from 'classnames';
 import { observer } from 'mobx-react-lite';
 import React from 'react';
 
+import { VIEW_CONTAINERS } from '@opensumi/ide-core-browser/lib/layout/view-id';
 import { generateCtxMenu, ICtxMenuRenderer } from '@opensumi/ide-core-browser/lib/menu/next';
 import { useInjectable } from '@opensumi/ide-core-browser/lib/react-hooks';
 import { IStatusBarService } from '@opensumi/ide-core-browser/lib/services';
 
 import { StatusBarItem } from './status-bar-item.view';
 import styles from './status-bar.module.less';
-
 
 export const StatusBarView = React.memo(
   observer(() => {
@@ -30,7 +30,12 @@ export const StatusBarView = React.memo(
     }, []);
 
     return (
-      <div className={styles.statusBar} style={{ backgroundColor }} onContextMenu={handleCtxMenu}>
+      <div
+        id={VIEW_CONTAINERS.STATUS_BAR}
+        className={styles.statusBar}
+        style={{ backgroundColor }}
+        onContextMenu={handleCtxMenu}
+      >
         <div className={cls(styles.area, styles.left)}>
           {statusBar.leftEntries.map((entry) => (
             <StatusBarItem key={entry.entryId} {...entry} />
